rearrange dashboard components into seperate files
